Add ability to bury layer
This lets us shove a layer at the beginning of the list of layers,
essentially burying it. This can be used to bury items under the ground
later.

Also adding something so that there will always be a floor present in
the cell, unless the cell is intentionally empty. Floor is probably
something we want no matter what, even if the other layer is a wall. I
ended up dealing with this in my game explicitly, but I think this makes
more sense here.

It also adds a function to tell if the cell is completely obstructed
with something that is meant to be never broken, the outer wall.