More in the grid
Add another method to get a random coordinate that is not completely
obstructed (edge walls are not possible to get) and another to be able
to bury layers.

I also moved the rng to the struct state so that every time a new
number is generated it uses the same rng. I was finding that I kept
getting the same index when I was trying to bury hidden items.